requirements.txt
keras==2.8.0
numpy==1.22.3
pika==1.2.1
Pillow==9.1.0
receive==0.1.1
